Moved test_pure_tabulate to tests
